moving notfound to Pages folder stucture
