+++ webui/app-tracker/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./components/Pages/Home/Home";
 import SearchApplications from "./components/Pages/Search/SearchApplications";
 import AddApplication from "./components/Pages/AddApplication/AddApplication";
-import NotFound from "./components/NotFound";
+import NotFound from "./components/Pages/NotFound/NotFound";
 import Statistics from "./components/Pages/Statistics/Statistics";
 
 function App() {
