+++ webui/app-tracker/src/components/Pages/NotFound/NotFound.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Menu from "./Menu/Menu";
+import Menu from "../../Menu/Menu";
 
 export default function NotFound() {
   return (
